rename: infrastructure 패키지 추가 aws.s3 [AIR-128]
