+++ aircnc/src/main/java/com/gurudev/aircnc/domain/room/service/RoomPhotoServiceImpl.java
@@ -2,7 +2,7 @@ package com.gurudev.aircnc.domain.room.service;
 
 import com.gurudev.aircnc.domain.base.AttachedFile;
 import com.gurudev.aircnc.domain.room.entity.RoomPhoto;
-import com.gurudev.aircnc.infrastructure.s3.S3Client;
+import com.gurudev.aircnc.infrastructure.aws.s3.S3Client;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
