+++ aircnc/src/test/java/com/gurudev/aircnc/domain/room/service/RoomPhotoServiceImplTest.java
@@ -10,7 +10,7 @@ import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.PutObjectRequest;
 import com.gurudev.aircnc.domain.base.AttachedFile;
 import com.gurudev.aircnc.domain.room.entity.RoomPhoto;
-import com.gurudev.aircnc.infrastructure.s3.S3Client;
+import com.gurudev.aircnc.infrastructure.aws.s3.S3Client;
 import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStream;
